requirements.txt
pdfplumber==0.10.3
faiss-cpu==1.8.0
sentence-transformers==2.7.0
requests==2.31.0
